🐞 fix(Fix middleware): Fix and refactor
Created useTypedSelector ;
Created useAppDispatch;
deleted IReducers and added typeof dispatch and store ;
