--- src/store/interfaces/IReducers.ts
@@ -1,7 +1,0 @@
-import { IContributor, IUser } from './IDataUser';
-import { ILoadings } from './ILoadings';
-export interface RootState {
-  userReducer: IUser;
-  reviewerReducer: IContributor;
-  loadingsReducer: ILoadings;
-}
